+++ src/MCDRpost/mcdrpost/manager/version_manager.py
@@ -6,6 +6,7 @@ from mcdreforged import PluginServerInterface
 from mcdrpost.data_structure import Item
 from mcdrpost.environment import Environment
 from mcdrpost.version_handler.abstract_version_handler import AbstractVersionHandler
+from mcdrpost.version_handler.sound_player.abstract_sound_player import AbstractSoundPlayer
 
 Checker = Callable[[Environment], bool]
 
@@ -91,3 +92,7 @@ class VersionManager:
         if self._handler is None:
             raise RuntimeError("version handler is not initialized")
         return self._handler.get_offhand_item(player)
+
+    @property
+    def play_sound(self) -> AbstractSoundPlayer:
+        return self._handler.play_sound