+++ src/MCDRpost/mcdrpost/utils/__init__.py
@@ -11,10 +11,13 @@ def get_formatted_time() -> str:
     return time.strftime('%Y-%m-%d %H:%M:%S', time.localtime())
 
 
-__Node = TypeVar("__Node", bound=AbstractNode)
+# 其实想用 3.12 的泛型语法的
+# 但是还是要保证兼容性
+# TODO: 此注解应该在 MCDR 产生放弃 3.10 版本的更新时替换成 3.12 的新泛型语法
+__NodeType = TypeVar("__NodeType", bound=AbstractNode)
 
 
-def add_requirements(node: __Node, permission: int, require_player: bool = False) -> __Node:
+def add_requirements(node: __NodeType, permission: int, require_player: bool = False) -> __NodeType:
     def require_callback(src: CommandSource) -> bool:
         if require_player and not src.is_player:
             return False
@@ -22,14 +25,14 @@ def add_requirements(node: __Node, permission: int, require_player: bool = False
             return False
         return True
 
-    def on_error(src: CommandSource):
+    def on_require_not_met(src: CommandSource):
         if require_player and not src.is_player:
             src.reply(TranslationKeys.only_for_player.tr())
             return
         src.reply(TranslationKeys.no_permission.tr())
 
     node.requires(require_callback).on_error(
-        RequirementNotMet, on_error, handled=True
+        RequirementNotMet, on_require_not_met, handled=True
     )
 
     return node